ml/requirements.txt
torch
torchvision
huggingface-hub
pyyaml
tqdm